data-table/requirements.txt
google-generativeai==0.8.3
anthropic==0.40.0
pydantic==2.10.3
python-dotenv==1.0.1
loguru==0.7.2
pandas==2.2.3
openpyxl==3.1.5
